workaround for KafkaInputFormat not passing through value timestamp
Using KafkaInputFormat it is currently not possible to access the
timestamp of the InputRow returned by the value InputFormat unless it
is present as a separate field in the InputRow.

This adds a workaround to inject the value timestamp as the __time field
so it can be accessed via the timestampSpec.